+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ## 7.2.0-beta.4
 
+- perf: Scope sync to SentryCrash (#1193)
 - fix: Add viewAppearing to UIViewController spans (#1202)
 
 ## 7.2.0-beta.3